aplusarchitects/requirements.txt
amqp==5.1.1
asgiref==3.5.1
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.4
django-bootstrap4==22.1
django-celery-results==2.3.1
django-filter==21.1
django-render-block==0.9.1
django-templated-email==3.0.0
django-widget-tweaks==1.4.12
Faker==13.11.1
kombu==5.2.4
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
text-unidecode==1.3
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tzdata==2022.1
vine @ file:///Users/ktietz/demo/mc3/conda-bld/vine_1630579390042/work
wcwidth==0.2.5
wincertstore==0.2
